Error handle for partial
Uncaught bug where the logic for partial masks would cause "?s" to overwrite several other characters when creating partial masks. Because this only affects when "s" is used with other characters and the rest work together an exception was added where a blank line would be returned instead of a malformed mask.